[Modify #1] rename driver header
