+++ src/eeprom/driver.c
@@ -1,11 +1,8 @@
-//
-// ドライバ
-//
+#include "driver.h"
+
 #include <eeprom.pio.h>
 #include <hardware/pio.h>
 #include <pico/stdlib.h>
-
-#include "eeprom_driver.h"
 
 void eepromInitStateMachine(const EEPROM* eeprom) {
     uint offset = pio_add_program(eeprom->pio, &eeprom_program);
